+++ src/main/java/com/niladri/MoneyManagement/Models/Spend_Transaction.java
@@ -20,6 +20,6 @@ public class Spend_Transaction {
     public Date date;
     public String category;
     public String description;
-    public Double amount;
+    public Integer amount;
     public String user;
 }
